Not cleaning m2 repo anymore, only cleaning dk.dk artifacts

+++ Jenkinsfile
@@ -52,7 +52,7 @@ pipeline {
             steps {
                 cleanWs()
                 sh """
-                    rm -r '/home/isworker/.dataio_repo/dk/dbc'
+                    test -d '/home/isworker/.dataio_repo/dk/dbc' && rm -r '/home/isworker/.dataio_repo/dk/dbc'
                 """
                 checkout scm
             }
